project/requirements.txt
Flask
Flask-Session
requests
wikipedia
python-dotenv
pandas